% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttb.R
\name{ttb_search}
\alias{ttb_search}
\title{Search TTB datasets by keyword}
\usage{
ttb_search(query)
}
\arguments{
\item{query}{Character string to search in dataset names and descriptions}
}
\value{
tibble: matching datasets
}
\description{
Search TTB datasets by keyword
}
